requirements.txt
argparse==1.2.1
django-filebrowser==3.5.2
django-grappelli==2.4.7
wsgiref==0.1.2
dj-database-url==0.2.2
dj-static==0.0.5
django-registration==1.0
django-toolbelt==0.0.1
gunicorn==18.0
html2text==3.200.3
